Refactor to share handling between validator and metrics providers. Implement metrics provider.
